Rename widgets.hpp to aspect_frame_bg.hpp
